Enable rdsModel directive
Add unit tests to ensure parity on overlap with model directive
Fixes some issues with orchestration
